Refactor: Extract WidgetEditor to its own file (#553)
This is PR 10 in a series of PRs that restore some editor functionality that I [deleted](https://github.com/Khan/webapp/pull/3144) back in January of 2022. It obviously isn't used often, but it's [still needed](https://khanacademy.slack.com/archives/C6NS8DWJU/p1683664865635869)!

Previous PRs:
  * #536
  * #537
  * #539
  * #540
  * #542
  * #543
  * #544
  * #551

In this PR I extract WidgetEditor to its own file (including moving SectionControlButton into components/)

Issue: LC-813

## Test plan:

Start storybook and do some exploratory testing with the Interactive Graph editor demo.

Author: jeremywiebe

Reviewers: jeremywiebe, jaredly

Required Reviewers:

Approved By: jaredly

Checks: ✅ finish_coverage, ✅ Publish npm snapshot (ubuntu-latest, 16.x), ✅ Extract i18n strings (ubuntu-latest, 16.x), ✅ Check builds for changes in size (ubuntu-latest, 16.x), ✅ Lint, Typecheck, and Test (ubuntu-latest, 16.x), ✅ Jest Coverage (ubuntu-latest, 16.x), ✅ Cypress Coverage (ubuntu-latest, 16.x), ✅ Check for .changeset file (ubuntu-latest, 16.x), ✅ gerald

Pull Request URL: https://github.com/Khan/perseus/pull/553
