+++ packages/perseus-editor/src/components/section-control-button.tsx
